+++ src/main/java/com/connectCare/connectCareApi/services/impl/DisponibilidadeServiceImpl.java
@@ -1,6 +1,7 @@
 package com.connectCare.connectCareApi.services.impl;
 
 import com.connectCare.connectCareApi.exceptions.DisponibilidadeNaoEncontradaException;
+import com.connectCare.connectCareApi.exceptions.NaoAutorizadoException;
 import com.connectCare.connectCareApi.exceptions.NenhumRegistroEncontradoException;
 import com.connectCare.connectCareApi.exceptions.OperacaoBancoDeDadosException;
 import com.connectCare.connectCareApi.models.dtos.GetPorIntervaloDataDTO;
@@ -8,13 +9,16 @@ import com.connectCare.connectCareApi.models.entities.Disponibilidade;
 import com.connectCare.connectCareApi.models.entities.Medico;
 import com.connectCare.connectCareApi.repositories.DisponibilidadeRepository;
 import com.connectCare.connectCareApi.services.GenericService;
+import com.connectCare.connectCareApi.utils.UsuarioAutenticado;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class DisponibilidadeServiceImpl implements GenericService<Disponibilidade> {
@@ -28,6 +32,12 @@ public class DisponibilidadeServiceImpl implements GenericService<Disponibilidad
 	@Override
 	public Disponibilidade create(Disponibilidade disponibilidade) {
 		Medico medicoEncontrado = medicoService.getById(disponibilidade.getMedico().getId());
+
+		//Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao médico
+		if(!Objects.equals(medicoEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+			throw new NaoAutorizadoException();
+		}
+
 		disponibilidade.setMedico(medicoEncontrado);
 		return repository.save(disponibilidade);
 	}
@@ -53,6 +63,10 @@ public class DisponibilidadeServiceImpl implements GenericService<Disponibilidad
 
 	@Override
 	public List<Disponibilidade> getAll() {
+		//Verificando se o atual usuário é ADMIN
+		if(!UsuarioAutenticado.getUsuarioAutenticado().getAuthorities().contains(new SimpleGrantedAuthority("ADMIN"))){
+			throw new NaoAutorizadoException();
+		}
 		List<Disponibilidade> disponibilidadesEncontradas = repository.findAll();
 
 		if(disponibilidadesEncontradas.isEmpty()) throw new NenhumRegistroEncontradoException("Disponibilidade");
@@ -63,6 +77,13 @@ public class DisponibilidadeServiceImpl implements GenericService<Disponibilidad
 	public Disponibilidade update(Disponibilidade disponibilidade) {
 		try{
 			Disponibilidade disponibilidadeEncontrada = repository.getReferenceById(disponibilidade.getId());
+
+			Medico medicoDisponibilidade = disponibilidadeEncontrada.getMedico();
+
+			//Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao médico
+			if(!Objects.equals(medicoDisponibilidade.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+				throw new NaoAutorizadoException();
+			}
 
 			disponibilidadeEncontrada.setAgendado(disponibilidade.isAgendado());
 			disponibilidadeEncontrada.setDataDisponivel(disponibilidade.getDataDisponivel());
@@ -80,6 +101,14 @@ public class DisponibilidadeServiceImpl implements GenericService<Disponibilidad
 	public void delete(Integer id) {
 		try{
 			Disponibilidade disponibilidadeEncontrada = repository.findById(id).orElseThrow(() -> new DisponibilidadeNaoEncontradaException(id));
+
+			Medico medicoDisponibilidade = disponibilidadeEncontrada.getMedico();
+
+			//Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao médico
+			if(!Objects.equals(medicoDisponibilidade.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+				throw new NaoAutorizadoException();
+			}
+
 			repository.delete(disponibilidadeEncontrada);
 		}catch(DataIntegrityViolationException e) {
 			throw new OperacaoBancoDeDadosException("Não foi possível excluir, pois essa disponibilidade está relacionada com alguma consulta.");
